+++ HISTORY.md
@@ -5,6 +5,9 @@
 - Moved repo to [`scarlatti` organization](https://github.com/scarlatti)
 - Fix protected range bug: specify emails of editors (which is only the owner)
   (https://github.com/scarlatti/roseingrave/issues/13)
+- Add text wrapping and disable automatic number formatting for any cells with
+  user input (https://github.com/scarlatti/roseingrave/issues/16
+  https://github.com/scarlatti/roseingrave/issues/18)
 
 ## v0.10.1 (2023-01-23)